+++ src/main/java/com/mraof/minestuck/blockentity/machine/TotemLatheBlockEntity.java
@@ -144,10 +144,7 @@ public class TotemLatheBlockEntity extends BlockEntity
 		BlockState state = level.getBlockState(pos);
 		
 		if(isValidDowelRod(state, facing))
-		{
 			level.removeBlock(pos, false);
-			setActiveRod(false);
-		}
 	}
 	
 	public boolean setDowel(ItemStack dowelStack)
@@ -168,7 +165,6 @@ public class TotemLatheBlockEntity extends BlockEntity
 				.setValue(TotemLatheBlock.DowelRod.DOWEL, EnumDowelType.getForDowel(dowelStack));
 		
 		level.setBlockAndUpdate(pos, newState);
-		setActiveRod(true);
 		
 		BlockEntity be = level.getBlockEntity(pos);
 		if(be instanceof ItemStackBlockEntity beItem)
@@ -210,22 +206,6 @@ public class TotemLatheBlockEntity extends BlockEntity
 			return true;
 		} else
 			return false;
-	}
-	
-	private void setActiveRod(boolean active)
-	{
-		Objects.requireNonNull(this.level);
-		Direction facing = this.getFacing();
-		
-		BlockPos rodPos = MSBlocks.TOTEM_LATHE.getRodPos(getBlockPos(), getBlockState());
-		BlockState rodState = this.level.getBlockState(rodPos);
-		if(rodState.is(MSBlocks.TOTEM_LATHE.ROD.get()) && rodState.getValue(TotemLatheBlock.FACING) == facing)
-			this.level.setBlockAndUpdate(rodPos, rodState.setValue(TotemLatheBlock.Rod.ACTIVE, active));
-		
-		BlockPos wheelPos = MSBlocks.TOTEM_LATHE.getWheelPos(getBlockPos(), getBlockState());
-		BlockState wheelState = this.level.getBlockState(wheelPos);
-		if(wheelState.is(MSBlocks.TOTEM_LATHE.WHEEL.get()) && wheelState.getValue(TotemLatheBlock.FACING) == facing)
-			this.level.setBlockAndUpdate(wheelPos, wheelState.setValue(TotemLatheBlock.Rod.ACTIVE, active));
 	}
 	
 	public ItemStack getDowel()